Moving utils to utils
