+++ data_adapter_oemof/utils.py
@@ -53,3 +53,37 @@ def convert_mixed_types_to_same_length(column):
         ]
     else:
         return column
+
+def divide_two_lists(dividend, divisor):
+    """
+    Divides two lists returns quotient, returns 0 if divisor is 0
+
+    Lists must be same length
+
+    Parameters
+    ----------
+    dividend
+    divisor
+
+    Returns divided list
+    -------
+
+    """
+    return [i / j if j != 0 else 0 for i, j in zip(dividend, divisor)]
+
+def multiply_two_lists(l1, l2):
+    """
+    Multiplies two lists
+
+    Lists must be same length
+
+    Parameters
+    ----------
+    l1
+    l2
+
+    Returns divided list
+    -------
+
+    """
+    return [i * j for i, j in zip(l1, l2)]
